+++ wnsm-smartmeter/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.2.2](https://github.com/ZakiZtraki/haos-addon-wnsm-dev/compare/v0.2.1...v0.2.2) (2025-05-29)
+
+
+### Bug Fixes
+
+* Fix parameter name issue in bewegungsdaten method call
+* Add fallback for different vienna-smartmeter library versions
+
 ### [0.2.1](https://github.com/ZakiZtraki/haos-addon-wnsm-dev/compare/v0.1.27...v0.2.1) (2025-05-29)